fix OWVibe module naming
